Fix issue when correction includes character *after* word
If you correct "anotherr test", one of the later suggestions is "another
retest", which includes the space after the misspelled word in the
correction. Make a basic fix for this, which is to include the character
after the correction in vim_you_autocorrect_before_correction and
vim_you_autocorrect_after_correction, and account for this by trimming
the highlighting by one character if it's a space.

This is obviously hacky, and still highlights other non-space
non-keyword characters.

A later commit will instead set vim_you_autocorrect_before_correction
and vim_you_autocorrect_after_correction more dynamically, based on the
content of the line before and after the correction.